Challenge order

+++ DonateChallenger/Services/ChallengeCatalog/ChallengeCatalog.API/Data/Entities/ChallengeEntity.cs
@@ -10,7 +10,7 @@ public class ChallengeEntity
 
     public string Description { get; set; } = null!;
 
-    public decimal DonatePrice { get; set; }
+    public double DonatePrice { get; set; }
 
     public string DonateFrom { get; set; } = null!;
